+++ content/footer.php
@@ -2,5 +2,6 @@
 		<footer>
 		
 		</footer>
+		<?php getScript('script.js'); ?>
 	</body>
 </html>